requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.12.1
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
fastapi==0.128.8
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-anthropic==0.3.22
langchain-core==0.3.84
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.0.2
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.6
pyarrow==21.0.0
pydantic==2.13.1
pydantic_core==2.46.1
pydeck==0.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
six==1.17.0
slack_bolt==1.28.0
slack_sdk==3.41.0
smmap==5.0.3
sniffio==1.3.1
starlette==0.49.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
xxhash==3.6.0
zstandard==0.25.0
